feat: implement a mock virtual object manager to support unit tests outside SwingSet
